requirements.txt
telethon>=1.28.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0
Pillow>=10.0.0
